Add functionality for user to update DOM with search results
